Move Model class to schemas/models
